add(many)

+++ app/Http/Controllers/TblUserController.php
@@ -3,7 +3,9 @@
 namespace App\Http\Controllers;
 
 use App\Models\TblUser;
+use Hash;
 use Illuminate\Http\Request;
+use RealRashid\SweetAlert\Facades\Alert;
 
 class TblUserController extends Controller
 {
@@ -26,17 +28,53 @@ class TblUserController extends Controller
         $data = ['user' => TblUser::all()];
         return view('pengguna.tambah', $data);
     }
-
-    public function edit()
+    
+    public function simpan(Request $request)
     {
-        $data = ['user'=>TblUser::all()];
-        return view('pengguna.edit', $data);
+        $data = $request->validate([
+            'username' => 'required',
+            'password' => 'required',
+            'role' => 'required',
+        ]);
+        $user = new TblUser($data);
+        $user->password = Hash::make($data['password']);
+        $user->save();
+        Alert::success('Data berhasil ditambah');
+        return redirect()->to('/dashboard/manage-user');
     }
 
-    public function delete($id)
+
+    public function edit($id)
     {
         $user = TblUser::find($id);
-        $user->delete();
-        return redirect()->route('pengguna.index')->with('success', 'Data Berhasil Dihapus');
+        
+        if (!$user) {
+            return redirect()->route('pengguna.index')->with('error', 'Pengguna tidak ditemukan');
+        }
+        
+        return view('pengguna.edit', compact('user'));
     }
+    public function update(Request $request, $id)
+    {
+        $data = $request->validate([
+            'username' => ['required', 'max:40'],
+            'role' => ['required']
+        ]);
+
+        $user = TblUser::query()->find($id);
+        $user -> fill($data);
+        $user -> save();
+
+        return redirect('/dashboard/manage-user')->with('success', 'User berhasil di update');
+    }
+    
+    public function destroy($id)
+    {
+       $user = TblUser::find($id);
+       if($user){
+            $user->delete();
+        }
+       return redirect()->route('pengguna.index')->with('success', 'Data berhasil dihapus');
+    }
+
 }
